eslint fix

+++ .eslintrc.js
@@ -156,7 +156,7 @@ module.exports = {
       }
     }],
     'padded-blocks': [2, 'never'],
-    'quotes': [2, 'single', {
+    'quotes': [2, 'double', {
       'avoidEscape': true,
       'allowTemplateLiterals': true
     }],
